Fix sony arw v1 Fix msbpump (getBit()) Remove some unsafe code
Former-commit-id: ac715fcd1eec442ca49abbd9c2d24a24c46ec907 [formerly d4b6256a5b6aeb581878866fc4f7c42b515d1e0c] [formerly 9076e1c59d0f2fa36361dbf978655c081f1c633b [formerly d893be1f95936a100aead5c684740c2c1bf64b85] [formerly d5a1cd0337b5feebe647ae0c853b3d48c40706d2] [formerly 2ff70062e2c98e9cf323fb174f480d7c2abd498f [formerly 36f0b86c8300b9d39dfc4ba926b270a9e124e461]] [formerly 9076e1c59d0f2fa36361dbf978655c081f1c633b [formerly d893be1f95936a100aead5c684740c2c1bf64b85] [formerly d5a1cd0337b5feebe647ae0c853b3d48c40706d2] [formerly 2ff70062e2c98e9cf323fb174f480d7c2abd498f [formerly 36f0b86c8300b9d39dfc4ba926b270a9e124e461]] [formerly 271832da01230ff80c477aae1c90b8a881c23e71]]]
Former-commit-id: 587fe5dea58c1b6ee27251a7846f733112cca1e6
Former-commit-id: 7c4f1dc7c0671d1123ec006b1ae6564dcadd4b4f
Former-commit-id: fb76a7e8475236a493931102cbb18ecf6903e0cb
Former-commit-id: ea21108ef1a4d966076a9c66f256094aab3e2ba4
Former-commit-id: 842a4360aa5c1b99daa6f13972d6353b0c5e8174 [formerly 6352562b582a78f1faf21a7bd288ce2f4d3dace0]
Former-commit-id: df48b6e6a33bda25585df5a224178861b0cae28c
Former-commit-id: 54a2af8f742d335290f3ebf55d721080a1cdc122

+++ Source/RawParser/Model/Parser/Decoder/Decompressor/BitPump/BitPump.cs
@@ -11,13 +11,13 @@ namespace RawNet.Decoder.Decompressor
         protected static int BITS_PER_LONG = (8 * sizeof(UInt32));
         protected static int BITS_PER_LONG_LONG = (8 * sizeof(UInt64));
         protected int left;
-        protected ulong current;
+        protected int current;
 
         public abstract int Offset { get; set; }
         public abstract void Fill();
 
-        public abstract uint PeekBit();
-        public abstract uint GetBit();
+        public abstract int PeekBit();
+        public abstract int GetBit();
         public abstract void SkipBits(int nbits);
         public abstract uint GetBits(int nbits);
         public abstract uint PeekBits(int nbits);
